refactor: clean up unused interfaces and improve dependency handling
- Removed unused User, Comment, and Report interfaces from ReportList and ReportViewer components.
- Updated useEffect dependencies in ReportPanel to include fetchReports.
- Changed ApiResponse and PaginatedResponse generic types from 'any' to 'unknown' for better type safety.
- Adjusted import paths for NextAuth and ApiResponse in relevant files.
