+++ lib/auth.ts
@@ -1,4 +1,4 @@
-import { AuthOptions } from "next-auth";
+import { AuthOptions } from "next-auth/";
 import CredentialsProvider from "next-auth/providers/credentials";
 import { compare } from "bcryptjs";
 import prisma from "./prisma";
